EncryptView: Implement Scrypt encryption
Take password and plaintext input, create an Age ScryptRecipient and
pass it through Age.encrypt, finally creating an armoured output into a
text box that copies to the clipboard on tap.

A share sheet is also added to easily send the armoured output to
another app or save to a file.