Settings: Add CreateKeyView sheet
Pop up a sheet when the "Create key" button is pressed, that asks you to
give the key a name (for user identification when using it), and
displaying the public and private keys after generation.

The name is checked for uniqueness to ensure things can't get muddled
up.

Generated keys are not currently persisted.