CreateKeyView: Move Done button to a NavigationStack
I realised I can make this look a bit more "usual" by putting the Done
button in the top bar of the sheet by using a toolbar, which requires a
Navigation{View,Stack}.

Using NavigationStack requires iOS 16, which could be limiting, but
there are other components in AgeKit which require newer macOS/iOS
versions so it's not a huge limitation. And iOS 17 is coming out later
this year anyway.