Decrease logo size, remove padding
Make the logo slightly smaller as each tab can scroll under it when
there's a lot of content (e.g. after encrypting and presenting the PEM),
and the logo takes a fair bit of space.

Also remove the padding around the main content as it doesn't look great
in light mode: the forms cause a grey background with a straight white
line separating the tabs at the bottom.