Save to and load keys from Keychain
Until now the generated keys only lasted as long as the app was running.
On relaunch they would be lost.

Now there is a rudimentary Keychain handler to save keys as a generic
password, storing the public key as the "service" name so we can access
it without having to regenerate it from AgeKit; and key names and public
keys can be fetched from the Keychain again and stored into an array for
display in the settings view.

Tapping on a key in Settings shows an alert with the public key. This
isn't very useful for now, but it gives a rough idea how this can work.